* Added PCH support in contrib plugins.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@1028 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/devpak_plugin/devpakinstaller.cpp
@@ -133,7 +133,7 @@ bool DevPakInstaller::Uninstall(const wxString& entry)
     for (unsigned int i = 0; i < pathlist.GetCount(); ++i)
     {
         wxString path = pathlist[i];
-        size_t pos = path.Find(_T('/'), true);
+        int pos = path.Find(_T('/'), true);
         while (pos != wxNOT_FOUND)
         {
             wxRmdir(path);
